pyproject.toml
[project]
name = "pyf.aggregator"
version = "1.0.0a1.dev0"
description = "Python Package Filter Aggregator"
keywords = ["PyPI"]
readme = "README.md"
authors = [
]
requires-python = ">=3.12"
license = { text = "Apache-2.0" }

classifiers = [
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Development Status :: 3 - Alpha",
]
dependencies = [
    "lxml",
    "PyYAML",
    "requests",
    "PyGithub",
    "typesense",
    "python-dotenv",
    "docutils",
    "readme-renderer[md]",
    "feedparser",
    "celery[redis]",
]

[project.urls]
Source = "https://github.com/collective/pyf.aggregator"



[project.optional-dependencies]
test = [
    "pytest-cov",
    "pytest",
    "responses",
    "webtest",
    "ruff",
]
typecheck = [
    "pytest-stub",
]

[project.scripts]
pyfaggregator = "pyf.aggregator.main:main"
pyfnpm = "pyf.aggregator.npm_main:main"
pyfupdater = "pyf.aggregator.typesense_util:main"
pyfgithub = "pyf.aggregator.enrichers.github:main"
pyfdownloads = "pyf.aggregator.enrichers.downloads:main"
pyfhealth = "pyf.aggregator.enrichers.health_calculator:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/pyf"]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
addopts = "-s --cov=pyf.aggregator --cov-report term --cov-report html"
markers = [
    "integration: mark test as integration test requiring Typesense",
]

[tool.isort]
profile = "plone"
src_paths = ["src", "tests"]

[tool.black]
target-version = ["py312"]
include = "src,tests"

[tool.codespell]
ignore-words-list = "discreet,"

[tool.check-manifest]
ignore = [
    ".coverage",
    ".flake8",
    ".pre-commit-config.yaml",
    ".vscode/*",
    "CHANGELOG.md",
    "htmlcov/*",
    "LICENSE.md",
    "Makefile",
    "mx.ini",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "responses>=0.25.8",
]
